Pawn hash table
Bench: 6743217
